feat: Added the Functionality to chnage the preffered audio device
