+++ lib/controllers/audio_device_controller.dart
@@ -6,9 +6,7 @@ import 'package:get/get.dart';
 import 'package:resonate/models/audio_device.dart';
 
 class AudioDeviceController extends GetxController {
-  final RxList<AudioDevice> audioInputDevices = <AudioDevice>[].obs;
   final RxList<AudioDevice> audioOutputDevices = <AudioDevice>[].obs;
-  final Rx<AudioDevice?> selectedAudioInput = Rx<AudioDevice?>(null);
   final Rx<AudioDevice?> selectedAudioOutput = Rx<AudioDevice?>(null);
 
   Timer? _deviceEnumerationTimer;
@@ -32,26 +30,13 @@ class AudioDeviceController extends GetxController {
   Future<void> enumerateDevices() async {
     try {
       final devices = await webrtc.navigator.mediaDevices.enumerateDevices();
-      final audioDevices = devices
+      final outputs = devices
           .map((device) => AudioDevice.fromMediaDeviceInfo(device))
+          .where((d) => d.isAudioOutput)
           .toList();
-
-      final inputs = audioDevices.where((d) => d.isAudioInput).toList();
-      final outputs = audioDevices.where((d) => d.isAudioOutput).toList();
-
-      audioInputDevices.value = inputs;
       audioOutputDevices.value = outputs;
-      if (selectedAudioInput.value == null && inputs.isNotEmpty) {
-        selectedAudioInput.value = inputs.first;
-        log('Default audio input device: ${inputs.first.label}');
-      }
-      if (selectedAudioOutput.value == null && outputs.isNotEmpty) {
-        selectedAudioOutput.value = outputs.first;
-        log('Default audio output device: ${outputs.first.label}');
-      }
-      log(
-        'Enumerated ${inputs.length} input and ${outputs.length} output devices',
-      );
+      selectedAudioOutput.value ??= outputs.firstOrNull;
+      log('Enumerated ${outputs.length} output devices');
     } catch (e) {
       log('Error enumerating devices: $e');
     }
@@ -61,65 +46,14 @@ class AudioDeviceController extends GetxController {
     try {
       selectedAudioOutput.value = device;
       await webrtc.Helper.selectAudioOutput(device.deviceId);
-      await _autoSelectInputDevice(device);
       log('Selected audio output: ${device.label}');
     } catch (e) {
       log('Error selecting audio output: $e');
     }
   }
 
-  Future<void> _autoSelectInputDevice(AudioDevice outputDevice) async {
-    try {
-      AudioDevice? inputDevice;
-      if (outputDevice.groupId.isNotEmpty) {
-        inputDevice = audioInputDevices.firstWhereOrNull(
-          (input) => input.groupId == outputDevice.groupId,
-        );
-      }
-      if (inputDevice == null) {
-        final outputLabel = outputDevice.label.toLowerCase();
-        if (outputLabel.contains('bluetooth')) {
-          inputDevice = audioInputDevices.firstWhereOrNull(
-            (input) => input.label.toLowerCase().contains('bluetooth'),
-          );
-        } else if (outputLabel.contains('headset') ||
-            outputLabel.contains('headphone')) {
-          inputDevice = audioInputDevices.firstWhereOrNull(
-            (input) =>
-                input.label.toLowerCase().contains('headset') ||
-                input.label.toLowerCase().contains('headphone'),
-          );
-        } else if (outputLabel.contains('usb')) {
-          inputDevice = audioInputDevices.firstWhereOrNull(
-            (input) => input.label.toLowerCase().contains('usb'),
-          );
-        }
-      }
-      inputDevice ??= audioInputDevices.firstWhereOrNull(
-        (input) =>
-            input.label.toLowerCase().contains('built-in') ||
-            input.label.toLowerCase().contains('internal') ||
-            input.label.toLowerCase().contains('phone') ||
-            input.label.toLowerCase().contains('default'),
-      );
-      inputDevice ??= audioInputDevices.firstOrNull;
-      if (inputDevice != null) {
-        selectedAudioInput.value = inputDevice;
-        await webrtc.Helper.selectAudioInput(inputDevice.deviceId);
-        log(
-          'Auto-selected input device: ${inputDevice.label} for output: ${outputDevice.label}',
-        );
-      } else {
-        log('Warning: No input device found to pair with output device');
-      }
-    } catch (e) {
-      log('Error auto-selecting input device: $e');
-    }
-  }
-
   String getDeviceName(AudioDevice device) {
     final label = device.label.toLowerCase();
-
     if (label.contains('earpiece') || label.contains('receiver')) {
       return 'Phone Earpiece';
     } else if (label.contains('speaker')) {
@@ -133,13 +67,11 @@ class AudioDeviceController extends GetxController {
     } else if (label.contains('usb')) {
       return 'USB Audio';
     }
-
     return device.label.isNotEmpty ? device.label : 'Unknown Device';
   }
 
   String getDeviceIcon(AudioDevice device) {
     final label = device.label.toLowerCase();
-
     if (label.contains('earpiece') || label.contains('receiver')) {
       return 'phone';
     } else if (label.contains('bluetooth')) {
