+++ lib/controllers/friend_call_screen.dart
@@ -5,7 +5,7 @@ import 'package:resonate/controllers/friend_calling_controller.dart';
 import 'package:resonate/themes/theme_controller.dart';
 import 'package:resonate/utils/ui_sizes.dart';
 import 'package:resonate/views/widgets/room_header.dart';
-import 'package:resonate/views/widgets/audio_device_selector_dialog.dart';
+import 'package:resonate/views/widgets/audio_selector_dialog.dart';
 
 import 'package:resonate/l10n/app_localizations.dart';
