+++ lib/l10n/app_localizations_hi.dart
@@ -1346,4 +1346,19 @@ class AppLocalizationsHi extends AppLocalizations {
   @override
   String get noRecordingError =>
       'आपके पास कोई रिकॉर्डिंग नहीं है। लाइव चैप्टर रूम से बाहर निकलने के लिए, कृपया पहले रिकॉर्डिंग शुरू करें।';
+
+  @override
+  String get audioOutput => 'ऑडियो आउटपुट';
+
+  @override
+  String get selectPreferredSpeaker => 'अपना पसंदीदा स्पीकर चुनें';
+
+  @override
+  String get noAudioOutputDevices => 'कोई ऑडियो आउटपुट डिवाइस नहीं मिला';
+
+  @override
+  String get refresh => 'रीफ़्रेश करें';
+
+  @override
+  String get done => 'हो गया';
 }
