+++ lib/views/screens/room_screen.dart
@@ -10,7 +10,7 @@ import 'package:resonate/utils/ui_sizes.dart';
 import 'package:resonate/views/widgets/participant_block.dart';
 import 'package:resonate/views/widgets/room_app_bar.dart';
 import 'package:resonate/views/widgets/room_header.dart';
-import 'package:resonate/views/widgets/audio_device_selector_dialog.dart';
+import 'package:resonate/views/widgets/audio_selector_dialog.dart';
 
 class RoomScreen extends StatefulWidget {
   final AppwriteRoom room;
